requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
Pillow==10.4.0
transformers==4.44.2
torch==2.4.1
safetensors==0.4.5
